Cleanliness built-in post effect plugins(worked and unworked)
