+++ src/experimental-plugins/filler/package.json
@@ -1,0 +1,13 @@
+{
+    "name": "delir-plugin-filler",
+    "version": "0.0.1",
+    "author": "Ragg <ragg.devpr@gmail.com>",
+    "main": "index.js",
+    "engines": {
+        "delir": "0.0.0"
+    },
+    "delir": {
+        "name": "フィラー",
+        "type": "post-effect"
+    }
+}
