--- src/experimental-plugins/plane/package.json
@@ -1,15 +1,0 @@
-{
-    "name": "delir-plugin-plate",
-    "version": "0.0.1",
-    "author": "Ragg <ragg.devpr@gmail.com>",
-    "main": "index.js",
-    "engines": {
-        "delir": "0.0.x"
-    },
-    "delir": {
-        "feature": "CustomLayer",
-        "targetApi": {
-            "renderer": "0.0.x"
-        }
-    }
-}
